Add SetContentAndWidth method to Screen interface
tcell currently determines the width of a cell by checking
if the first rune in the cell is full width.  This works for
East Asian characters and emoji, but gives incorrect results
for some grapheme clusters.  For example, some Thai characters
are composed of multiple runes and occupy multiple cells,
but the sequence begins with a narrow rune.

Additionally, some callers may want to control rune width
based on user configuration, perhaps to configure handling
of ambiguous East Asian characters.  Currently, the caller
can set env vars to control the behavior of go-runewidth,
but it would be nice to control this more directly.

It's difficult for tcell to handle the conflicting requirements
of any application for determining cell width.  This commit
adds an additional method SetContentAndWidth to the Screen
interface so that applications that need more control
can set the width explicitly.